Start work on TableOperations.joinpartitions
With the new `Tables.partitions` functionality in the Tables.jl package,
it can be helpful for non-partitioned in-memory table structures to have
a way to treat the separate partitions as a single, long partition. The
SentinelArrays.jl package includes a `ChainedVector` type that allows
lazily "chaining" AbstractArrays together, treated as a single, long
vector. `TableOperations.joinpartitions(x)` takes any partitioned input
`x`, and appends the columns together via `ChainedVector`, producing a
`JoinedPartitions` object which itself, satisfies the `Tables.columns`
interface, i.e. you could do `df =
DataFrame(TableOperations.joinpartitions(x))` to concatenate all
partitions of the input and operate on them as a whole in a `DataFrame`.

Still need to add tests and docs.