Add TableOperations.makepartitions functionality
This adds a new function `TableOperations.makepartitions(x, nrows)` that
allows partitioning an input table into batches of `nrows` rows. It
works on any Tables.jl sources, regardless of column/row orientation.
For columns and AbstractVector inputs, we use "views" over the input for
efficiency, and for unknown-length input iterators, or when
`materializerows=true`, input rows will be copied/allocated to ensure
correctness. This is important for cases like `SQLite.Query`, where each
row is only valid on the current iteration, so the row needs to be fully
materialized, i.e. copied in order to preserve values.